tk1: Add syscall trigger address
The TKey implements syscalls using non-riscv-standard interrupts and
does not fit the RISC-V interrupt model. For the now we implement
syscall trigger address by directly modifying CPU registers from the
machine.

This doesn't add support for the PicoRV32 specific instructions
`picorv32_retirq_insn` and `picorv32_maskirq_insn` so that needs to be
handled in firmware.