Add scripts to debug a TKey app with QEMU over a network
These scripts can be used to debug for example a FIDO application
on Windows but the QEMU emulation for the app is run Linux.

The physical TKey should load the "Loopback" app from the
tillitis-key1 repo. The Loopback app should be modified to expose
the desired USB interface towards the host (e.g. Windows).

The tkey_to_udp_win.py script is started on the Windows host and
udp_to_qemu_linux.py is started on the Linux host together with
QEMU and the app to be debugged.

Windows can now use the USB interface of the physical Tkey for
communication but all execution is done on QEMU on Linux.